Always format qstates using parenthesis.
Attempting to format `(-a + bi)|0>` as `- a-bi|0>` was probably not a great idea.
So, for now, I am simply putting parenthesis around the amplitudes and always combining them with "+".

So, for example, `0.7071 |000> - 0.7071 |100>` will now be formatted as `(0.7071)|0> + (-0.7071)|1>

Also, I remove support for the 'spacedSign' flag. I am not using it anymore, and the lack of requests for bug fixes and features makes me think that no one is using this library besides me :-)

Hopefully this will help address:
https://github.com/davidbkemp/quantum-gate-playground/issues/1