Export jsqubits from index.js
So that you can write

    qubits = require('jsqubits').jsqubits

Hopefully this will restore backward compatibility with the previously published npm modules.

Continues to export Qubits as "default" which allows you to write:

    import Q from '../lib'

Bumping the package and bower versions to 1.1.0 as I suspect all the ES6 stuff is backward but not forward compatible.

A step towards publishing a new npm version
Git issue: #20