Remove reliance on Babel
To make this project easier to maintain I want to minimize the use of 3rd party tools as much as possible so I have less security updates.

Unfortunately eslint doesn't seem to understand static declarations in classes and so I am dropping it for now.  Will try to re-introduce it later.

I am not convinced that ES modules are the way to go.  To use them in the Node Repl you need to handle the promise returned by the import function.

To do:

- Try going back to CommonJS modules
- Revisit Browserify for packaging into something that can be used in browsers.