Improve instructions for running using Node
Probably don't need all of these extra isntructions. If user's can't figure it out for them selves then they are probably going to have bigger problems, but I guess it won't hurt.

Also, I figure it is safer for to explicitly list the npm files in package.json than to rely on .npmignore