Refactor constants and use github actions badge
- Put Complex constants back on Complex
- Put roundToZero and ALL on QSTATE
- Keep all the contants on the top level Qubits object.

This seems more sensible that having a constants object.  Ideally they would actually be static constants on the classes but Safari doesn't support them yet.

Unfortunately, they aren't really constants anymore. i.e. there is nothing stopping a client from changing their values.

roundToZero was never actually meant to be a constant. i.e. it is something that you should be able to configure.

Moving from Travis to Github actions.  Not sure what is going on with Travis.

I think this is backwards compatible with version 2.0.0